+++ src/main/java/br/com/gjnv/petshop/service/EnderecoService.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
 public class EnderecoService {
@@ -19,10 +20,10 @@ public class EnderecoService {
         enderecoRepository.save(endereco);
     }
 
-    public void removerEndereco(Long id) {
+    public void removerEndereco(Integer id) {
         enderecoRepository.deleteById(id);
     }
-    public Endereco atualizarEndereco(Long id, EnderecoDto enderecoDto) {
+    public Endereco atualizarEndereco(Integer id, EnderecoDto enderecoDto) {
         Optional<Endereco> enderecoOptional = enderecoRepository.findById(id);
 
         if (enderecoOptional.isPresent()) {
